Node movement logic has been moved out of board into the TFE "module" of C functions.
The grid is treated as an immutable list, passed in to moveNodesInDirection() and spawnNodesExcludingDirection(), with a copy returned. The functions do not hold any state, just manipulate the nodes and return the new configuration.

Dictionaries describing any node movement are passed back to TFEBoard indirectly from the movement functions. The board extracts the values and passes them to the scene for animating.