Restructured in preparation for sharing code
Introduced `TFEGameController` to handle overall game management,
including the score label. The board still communicates directly with the
scene, but only about node movements. The controller is responsible for
user input.

The `TFELabel` name was created to wrap either `NSTextField` or
`UILabel`, depending on platform. It exposes a `text` property.

A mostly-empty iOS target, with associated files, was added.