Established working cross-platform targets
`TFEGameController` now inherits from the wrapper `TFEViewController`,
which is a subclass of either `UIViewController` or `NSViewController`
depending on target platform. The app delegate on each platform sets the
controller as the window's content controller as appropriate.

The `keyDown:` method has been moved out to a category that will only be
compiled into the Mac target. Likewise, a `Touch` category accepts input
from `UISwipeGestureRecognizer`s in the iOS MainView nib.

The Mac target has a new MainView nib as well, which contains the views
that used to be directly in the window in MainMenu.xib.