Converted TFEGameController to Swift
`TFELabel` was also converted, and of course `TFEViewController`.

Some frustrating challenges with this, owing to Swift's insistence on
compile-time checks. See the new Notes.md file for details.

The two highlights were `NSViewController.viewWillAppear()` and
`UIViewController.viewWillAppear(_:)`, along with the differing
failablilty of the Swift version of `initWithNibName:bundle:` on the two
different platforms.

The creation of the controller's board required an IUO because the two
objects keep references to each other. `TFEBoard`'s relationship to the
controller should probably be changed when it is translated, either to
use callbacks or to have an Optional property instead of taking the
controller as a constructor parameter.