Converted TFEBoard to Swift
It is a struct, rather than a class. This necessitated a small change to
`TFEGameController`'s input handling. Since the board is a Swift struct,
it is not visible at all in ObjC. This is resolved by making the
controller call its own method in the main class body from the input
handling extensions, `userDidInput(_:)`, which then calls through to the
board.

In addition, the board struct mutates when a move is made. Therefore the
board property must be settable inside the extension. It isn't, because
the property is `private(set)`, but the setter is visible from
`userDidInput(_:)`.