Improved node animation appearance and code
If a node is not actually moving when forming a combination, it does not
wait to fade and shrink, and it is not given a move action at all. This
gives a nicer visual interaction with the node spawning on top of it.

Restructured/clarified code for spawn actions somewhat. In particular,
the "grow" and "bounce" actions are sequenced, rather than grouped with a
separate "wait" action on the bounce.