Continued WIP on converting Game Core
Engine functions are translated in TFE.swift, largely based on an ASCII-
based version of the game written in April. Further work needs to be done
to integrate them with `TFEBoard`. The interface has changed from the C
version because of the availability of tuples, but the use of `TFEMove`
complicates the integration slightly from the text version.

The grid modifying functions probably need to use `inout` rather than
directly returning the new grid.