Logic bug fixes on WIP converting to Swift
`TFENode` had an implicit member access, using `position` instead of
`destination` when it was told to move.

Test for possible movement in `TFEIsALoser()` was reversed.
`TFEMoveNodes(_:inDirection:)` was not actually updating the grid.
(Some variable names were improved.)

Set+RandomInt had an off-by-one that meant it never used the second
value.

The two GridSquarePositions structs had `down` and `up` swapped.