Prepared game core for packing into a module
A protocol `Tile` was added to the game core. The engine functions in
TFE.swift operate on this, rather than explicitly on `TFENode`s. This
will allow the core to be placed in its own module, which should in turn
help testability.

Across the project, the term "node" was replaced with "tile" to reflect
this as appropriate (including comments).

The functions in the engine, along with the (renamed) `TileMove` and
`SlidTile` types, have become generic as a consequence of the protocol's
introduction.